OpenSSLCertUtils::readCertsFromBuffer() to throw on malformed cert.
Summary:
Current behavior of OpenSSLCertUtils::readCertsFromBuffer() is that it stops
parsing as soon as it encounters the first error. We don't know if the error
is EOF or something else. Sometimes we want to reject the result
if the buffer is malformed instead of accepting a partially parsed result.
This diff makes the API throw when given malformed cert buffer.

Reviewed By: anirudhvr

Differential Revision: D10467792

fbshipit-source-id: 2c15266e5f00866dfaafe0a5ce88d24459e8b561