add a `detach` option to `Subprocess` to avoid waiting for the child
Summary:
Add a `Subprocess::Options::detach()` call.  This causes the subprocess to be
spawned in grandchild process that will be reparented to init.  The
intermediate child process exits immediately after spawning the grandchild,
allowing the parent process to `wait()` on it and destroy its `Subprocess`
object without leaving any zombie children.

Reviewed By: yfeldblum

Differential Revision: D10860274

fbshipit-source-id: ac27d77d0033e3c03e27dd3f758ab70fd2dd990f