ConcurrentHashMap: Temporarily disable using hazptr_obj_batch for bucket arrays.
Summary:
Completes the previous change to disable destruction order guarantee for keys and values to give some higher-level users of ConcurrentHashMap time to adapt to the destruction order guarantee.
To reenable the destruction order guarantee, uncomment the two calls to set_batch_tag that were commented out in this and the previous change.

Reviewed By: davidtgoldblatt

Differential Revision: D13141742

fbshipit-source-id: c49ad5b0a0f82da7b36e0fea294425ec2b681ab9